only test passwords for running docker instances

+++ scripts/startup/pg-auth-scan.sh
@@ -1,28 +1,34 @@
 #!/bin/bash
-#ip=$1
+
+# sanity check, then get ip
+ping icanhazip.com -c 1 &>/dev/null || exit 1
+if [[ ! -e /var/plexguide/basics.yes ]]; then
+  exit 1; fi
+ip=$(curl icanhazip.com)
+
 
 sonarr () {
 curl -m 5 -s $ip:8989 -o html \
   && grep -iq api html \
-  && echo "Sonarr Has No Password On $ip:8989"
+  && echo "WARNING: Sonarr Has No Password On $ip:8989" >> /var/plexguide/nopassword
 }
 
 radarr () {
 curl -m 5 -s $ip:7878 -o html \
   && grep -iq api html \
-  && echo "Radarr Has No Password On $ip:7878"
+  && echo "Radarr Has No Password On $ip:7878" >> /var/plexguide/nopassword
 }
 
 lidarr () {
 curl -m 5 -s $ip:8686 -o html \
   && grep -iq api html \
-  && echo "Lidarr Has No Password On $ip:8686"
+  && echo "Lidarr Has No Password On $ip:8686" >> /var/plexguide/nopassword
 }
 
 portainer () {
 curl -m 5 -s $ip:9000/api/users/admin/check -o html \
   && grep -iq '{"err":"User not found"}' html \
-  && echo "Portainer Has No Password On $ip:9000"
+  && echo "Portainer Has No Password On $ip:9000" >> /var/plexguide/nopassword
 }
 
 deluge () {
@@ -30,40 +36,34 @@ curl -d '{"method":"auth.login","params":["deluge"],"id":7}' \
   -H "Content-Type: application/json" \
    $ip:8112/json -m 5 -i -s -o html \
    && grep -iq 'Set-Cookie:' html \
-   && echo "Deluge Has Default Password: "deluge" On $ip:8112"
+   && echo "Deluge Has Default Password: "deluge" On $ip:8112" >> /var/plexguide/nopassword
 }
 
 rutorrent () {
 curl -i -m 5 -s $ip:8999 -o html \
   && grep -iq '200 OK' html \
-  && echo "ruTorrent Has No Password On $ip:8999"
+  && echo "ruTorrent Has No Password On $ip:8999" >> /var/plexguide/nopassword
+
 }
 
 nzbget () {
 curl -i -s -m 5 $ip:6789 -o html \
   && grep -q 'Auth-Type=http' html \
-  && echo "NZBget Has No Password On $ip:6789"
+  && echo "NZBget Has No Password On $ip:6789"  >> /var/plexguide/nopassword
 }
 
 sabnzbd () {
 curl -i -s -m 5 $ip:8080 -o html \
   && grep -q '200 OK' html \
-  && echo "SABNZBD Has No Password On $ip:8080"
+  && echo "SABNZBD Has No Password On $ip:8080"  >> /var/plexguide/nopassword
 }
 
-#for app in "portainer nzbget radarr sonarr"; do $app; done
-#for app in $(docker ps | awk '{print $NF}' | grep -v NAME); do $app; done
+# reset
+echo "" > /var/plexguide/nopassword
 
-for ip in $(cat plexopen); do
-  nzbget
-  rutorrent
-  deluge
-  sabnzbd
-  portainer
-  lidarr
-  sonarr
-  radarr
-done
+# only test passwords for docker containers that are running
+applist=$(docker ps | awk '{print $NF}' | grep -v NAME)
+for app in $applist; do $app; done
+
 
 exit 0
-
